slightly better project organization
